+++ crates/nostr-sdk/examples/negentropy.rs
@@ -22,9 +22,9 @@ async fn main() -> Result<()> {
     let my_items = Vec::new();
     let filter = Filter::new().author(public_key).limit(10);
     let opts = NegentropyOptions::default();
-    let Output { success, failed } = client.reconcile_with_items(filter, my_items, opts).await?;
-    println!("Success: {success:?}");
-    println!("Failed: {failed:?}");
+    let output = client.reconcile_with_items(filter, my_items, opts).await?;
+    println!("Success: {:?}", output.success);
+    println!("Failed: {:?}", output.failed);
 
     Ok(())
 }